+++ src/sess.rs
@@ -129,12 +129,7 @@ impl<'sess, 'ctx: 'sess> Session<'ctx> {
             cfg,
             manifest.package.name
         );
-        let src = match *cfg {
-            config::Dependency::Version(_) => DependencySource::Registry,
-            config::Dependency::Path(ref p) => DependencySource::Path(p.clone()),
-            config::Dependency::GitRevision(ref g, _)
-            | config::Dependency::GitVersion(ref g, _) => DependencySource::Git(g.clone()),
-        };
+        let src = DependencySource::from(cfg);
         self.deps
             .lock()
             .unwrap()
@@ -1381,6 +1376,17 @@ pub enum DependencySource {
     Path(PathBuf),
     /// The dependency is available at a git url.
     Git(String),
+}
+
+impl<'a> From<&'a config::Dependency> for DependencySource {
+    fn from(cfg: &'a config::Dependency) -> DependencySource {
+        match *cfg {
+            config::Dependency::Path(ref path) => DependencySource::Path(path.clone()),
+            config::Dependency::GitRevision(ref url, _) => DependencySource::Git(url.clone()),
+            config::Dependency::GitVersion(ref url, _) => DependencySource::Git(url.clone()),
+            config::Dependency::Version(_) => DependencySource::Registry,
+        }
+    }
 }
 
 impl fmt::Display for DependencySource {