+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `parents` command outputs more detail:
+  - source information in case multiple dependencies do not match
+  - information about an override if one is present
+
 ### Changed
 - Interpret all dependency names as lowercase to reduce ambiguity